requirements.txt
flask
prometheus-client
python-dotenv
boto3
